chore: refactor workspace package
