+++ packages/go/stbernard/analyzers/golang/golang.go
@@ -30,20 +30,6 @@ import (
 	"github.com/specterops/bloodhound/slicesext"
 )
 
-// InstallGolangCiLint runs go install for the currently supported golangci-lint version
-func InstallGolangCiLint(path string, env environment.Environment) error {
-	var (
-		command = "go"
-		args    = []string{"install", "github.com/golangci/golangci-lint/cmd/golangci-lint@v1.55.2"}
-	)
-
-	if err := cmdrunner.Run(command, args, path, env); err != nil {
-		return fmt.Errorf("golangci-lint install: %w", err)
-	}
-
-	return nil
-}
-
 // Run runs golangci-lint for all module paths passed to it
 //
 // This is a single runner that accepts the paths for all passed modules, rather than separate runs for each path
