+++ packages/go/stbernard/command/generate/generate.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/specterops/bloodhound/packages/go/stbernard/environment"
 	"github.com/specterops/bloodhound/packages/go/stbernard/workspace"
+	"github.com/specterops/bloodhound/packages/go/stbernard/workspace/golang"
 )
 
 const (
@@ -69,9 +70,9 @@ func (s *command) Parse(cmdIndex int) error {
 func (s *command) Run() error {
 	if paths, err := workspace.FindPaths(s.env); err != nil {
 		return fmt.Errorf("finding workspace root: %w", err)
-	} else if modPaths, err := workspace.ParseModulesAbsPaths(paths.Root); err != nil {
+	} else if modPaths, err := golang.ParseModulesAbsPaths(paths.Root); err != nil {
 		return fmt.Errorf("parsing module absolute paths: %w", err)
-	} else if err := workspace.WorkspaceGenerate(modPaths, s.env); err != nil {
+	} else if err := golang.WorkspaceGenerate(modPaths, s.env); err != nil {
 		return fmt.Errorf("generating code for workspace: %w", err)
 	} else {
 		return nil
