+++ packages/go/stbernard/workspace/golang/sync.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package workspace
+package golang
 
 import (
 	"errors"
